Adição I2C

+++ inc/interfaces.hpp
@@ -5,11 +5,11 @@
 #include "constantes.hpp"
 
 void mostrar_opcoes(WINDOW *window);
-void atualizar_menu(WINDOW *window, const int opcao_usuario, const int opcao_anterior, const int histerese);
+void atualizar_menu(WINDOW *window, const int opcao_usuario, const int opcao_anterior, const float histerese);
 void iniciar_saida(WINDOW *window, const int size_x);
 void iniciar_logs(WINDOW *window);
 void atualizar_logs(WINDOW *window, string dispositivo, const int status);
 void pegar_temperatura(WINDOW *window, const float TE, float *TR);
-void pegar_histerese(WINDOW *window, const int opcao_usuario, const int opcao_anterior, int *histerese);
+void pegar_histerese(WINDOW *window, const int opcao_usuario, const int opcao_anterior, float *histerese);
 
 #endif // INTERFACES_HPP
